cmake: Fix platform detection for Emscripten
ODE insists on defining a ODE_PLATFORM_* macro for each platform that's
hardcoded into the file, even though they're never used anywhere and
errors out on a platform that it does not explicitly recognize. Shut it
up by making Emscripten a recognized platform, rather than fixing the
spaghetti that is ODE's platform detection code properly, as we all have
better things to do with our time.

Signed-off-by: Alexander Shishkin <alex@ash.works>